* rename StreamOffset.value to StreamOffset.Offset * WIP; debug main

+++ Data.Test/Config.PC.cs
@@ -15,7 +15,7 @@ namespace DataBuildSystem
 		public string BigFileHdbExtension { get { return ".bfh"; } }
 		public string BigFileNodeExtension { get { return ".bfn"; } }
         public bool LittleEndian { get { return true; } }
-        public UInt32 FileAlignment { get { return 1024; } }
+        public Int64 FileAlignment { get { return 1024; } }
         public bool AllowDuplicateFiles { get { return false; } }
         public bool WriteAsync { get { return true; } }
         public UInt32 ReadBufferSize { get { return 8 * 1024 * 1024; } }
